requirements.txt
bokeh==3.1.1
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
seaborn==0.12.2
st_pages==0.4.1
streamlit==1.22.0
